T442: Some calculations are moved from StateWriterImpl to BlockDiff
