+++ src/main/scala/com/wavesplatform/state2/reader/StateReader.scala
@@ -43,6 +43,8 @@ trait StateReader {
   def findPreviousExchangeTxs(orderId: EqByteArray): Set[ExchangeTransaction]
 
   def isLeaseActive(leaseTx: LeaseTransaction): Boolean
+
+  def activeLeases(): Seq[ByteArray]
 }
 
 object StateReader {
